+++ lib/cloudcannon-jekyll/page-without-a-file.rb
@@ -1,6 +1,7 @@
 # frozen_string_literal: true
 
 module CloudCannonJekyll
+  # Utility class to help generate files with no source file
   class PageWithoutAFile < Jekyll::Page
     def read_yaml(*)
       @data ||= {}