+++ HISTORY.md
@@ -1,3 +1,15 @@
+# 0.3.0
+
+* Added drafts to collections in details
+* Added drafts and data to collections in config
+* Added category folder drafts and posts to collections in config
+* Added `_path` field to each collection definition in config
+* Removed some unused fields
+* Renamed static to static-pages in details and removed `robots.txt` and `sitemap.xml` exceptions
+* Add `url` to static-pages
+* Normalise `_path` in static-pages
+* Added max depth parameter for jsonify filter and increase it for array structures in config output
+
 # 0.2.2
 
 * Added JSON handling for integers in hash keys